requirements.txt
flask
keras
tensorflow==2.4.0
numpy
librosa==0.6.3
numba==0.48
tqdm
requests
matplotlib
pymongo==3.11.2
azure-storage-blob==2.1.0